+++ firebase-firestore/src/main/java/com/google/firebase/firestore/AggregateSource.java
@@ -14,6 +14,13 @@
 
 package com.google.firebase.firestore;
 
+/** Configures the behavior of {@code get()} calls on {@link AggregateQuery}. */
 public enum AggregateSource {
+  /**
+   * Reach to the Firestore backend and surface the result verbatim, that is no local documents or
+   * mutations in the SDK cache will be included in the surfaced result.
+   *
+   * <p>Requires client to be online.
+   */
   SERVER_DIRECT,
 }