+++ firebase-firestore/CHANGELOG.md
@@ -3,6 +3,8 @@ by opting into a release at
 [go/firebase-android-release](http:go/firebase-android-release) (Googlers only).
 
 # Unreleased
+- [feature] Add COUNT queries that go to server directly for the result, without applying local
+  documents and mutations.
 
 # 24.2.0
 - [feature] Added `TransactionOptions` to control how many times a transaction